Move CommonMetadata tests to separate module
